+++ lib/src/arg_parser_exception.dart
@@ -9,6 +9,14 @@ class ArgParserException extends FormatException {
   /// This will be empty if the error was on the root parser.
   final List<String> commands;
 
-  ArgParserException(super.message, [Iterable<String>? commands])
+  /// The name of the argument that was being parsed when the error was
+  /// discovered.
+  final String? argumentName;
+
+  ArgParserException(super.message,
+      [Iterable<String>? commands,
+      this.argumentName,
+      super.source,
+      super.offset])
       : commands = commands == null ? const [] : List.unmodifiable(commands);
 }