+++ lib/src/parser.dart
@@ -63,7 +63,8 @@ class Parser {
       // options so that commands can have option-like names.
       var command = _grammar.commands[_current];
       if (command != null) {
-        _validate(_rest.isEmpty, 'Cannot specify arguments before a command.');
+        _validate(_rest.isEmpty, 'Cannot specify arguments before a command.',
+            _current);
         var commandName = _args.removeFirst();
         var commandParser = Parser(commandName, command, _args, this, _rest);
 
@@ -71,7 +72,11 @@ class Parser {
           commandResults = commandParser.parse();
         } on ArgParserException catch (error) {
           throw ArgParserException(
-              error.message, [commandName, ...error.commands]);
+              error.message,
+              [commandName, ...error.commands],
+              error.argumentName,
+              error.source,
+              error.offset);
         }
 
         // All remaining arguments were passed to command so clear them here.
@@ -101,7 +106,7 @@ class Parser {
       // Check if an option is mandatory and was passed; if not, throw an
       // exception.
       if (option.mandatory && parsedOption == null) {
-        throw ArgParserException('Option $name is mandatory.');
+        throw ArgParserException('Option $name is mandatory.', null, name);
       }
 
       // ignore: avoid_dynamic_calls
@@ -118,11 +123,11 @@ class Parser {
   /// Pulls the value for [option] from the second argument in [_args].
   ///
   /// Validates that there is a valid value there.
-  void _readNextArgAsValue(Option option) {
+  void _readNextArgAsValue(Option option, String arg) {
     // Take the option argument from the next command line arg.
-    _validate(_args.isNotEmpty, 'Missing argument for "${option.name}".');
+    _validate(_args.isNotEmpty, 'Missing argument for "$arg".', arg);
 
-    _setOption(_results, option, _current);
+    _setOption(_results, option, _current, arg);
     _args.removeFirst();
   }
 
@@ -145,7 +150,8 @@ class Parser {
     var option = _grammar.findByAbbreviation(opt);
     if (option == null) {
       // Walk up to the parent command if possible.
-      _validate(_parent != null, 'Could not find an option or flag "-$opt".');
+      _validate(_parent != null, 'Could not find an option or flag "-$opt".',
+          '-$opt');
       return _parent!._handleSoloOption(opt);
     }
 
@@ -154,7 +160,7 @@ class Parser {
     if (option.isFlag) {
       _setFlag(_results, option, true);
     } else {
-      _readNextArgAsValue(option);
+      _readNextArgAsValue(option, '-$opt');
     }
 
     return true;
@@ -193,22 +199,23 @@ class Parser {
     var first = _grammar.findByAbbreviation(c);
     if (first == null) {
       // Walk up to the parent command if possible.
-      _validate(
-          _parent != null, 'Could not find an option with short name "-$c".');
+      _validate(_parent != null,
+          'Could not find an option with short name "-$c".', '-$c');
       return _parent!
           ._handleAbbreviation(lettersAndDigits, rest, innermostCommand);
     } else if (!first.isFlag) {
       // The first character is a non-flag option, so the rest must be the
       // value.
       var value = '${lettersAndDigits.substring(1)}$rest';
-      _setOption(_results, first, value);
+      _setOption(_results, first, value, '-$c');
     } else {
       // If we got some non-flag characters, then it must be a value, but
       // if we got here, it's a flag, which is wrong.
       _validate(
           rest == '',
           'Option "-$c" is a flag and cannot handle value '
-          '"${lettersAndDigits.substring(1)}$rest".');
+              '"${lettersAndDigits.substring(1)}$rest".',
+          '-$c');
 
       // Not an option, so all characters should be flags.
       // We use "innermostCommand" here so that if a parent command parses the
@@ -228,16 +235,16 @@ class Parser {
     var option = _grammar.findByAbbreviation(c);
     if (option == null) {
       // Walk up to the parent command if possible.
-      _validate(
-          _parent != null, 'Could not find an option with short name "-$c".');
+      _validate(_parent != null,
+          'Could not find an option with short name "-$c".', '-$c');
       _parent!._parseShortFlag(c);
       return;
     }
 
     // In a list of short options, only the first can be a non-flag. If
     // we get here we've checked that already.
-    _validate(
-        option.isFlag, 'Option "-$c" must be a flag to be in a collapsed "-".');
+    _validate(option.isFlag,
+        'Option "-$c" must be a flag to be in a collapsed "-".', '-$c');
 
     _setFlag(_results, option, true);
   }
@@ -269,16 +276,16 @@ class Parser {
     if (option != null) {
       _args.removeFirst();
       if (option.isFlag) {
-        _validate(
-            value == null, 'Flag option "$name" should not be given a value.');
+        _validate(value == null,
+            'Flag option "--$name" should not be given a value.', '--$name');
 
         _setFlag(_results, option, true);
       } else if (value != null) {
         // We have a value like --foo=bar.
-        _setOption(_results, option, value);
+        _setOption(_results, option, value, '--$name');
       } else {
         // Option like --foo, so look for the value as the next arg.
-        _readNextArgAsValue(option);
+        _readNextArgAsValue(option, '--$name');
       }
     } else if (name.startsWith('no-')) {
       // See if it's a negated flag.
@@ -286,18 +293,22 @@ class Parser {
       option = _grammar.findByNameOrAlias(positiveName);
       if (option == null) {
         // Walk up to the parent command if possible.
-        _validate(_parent != null, 'Could not find an option named "$name".');
+        _validate(_parent != null, 'Could not find an option named "--$name".',
+            '--$name');
         return _parent!._handleLongOption(name, value);
       }
 
       _args.removeFirst();
-      _validate(option.isFlag, 'Cannot negate non-flag option "$name".');
-      _validate(option.negatable!, 'Cannot negate option "$name".');
+      _validate(
+          option.isFlag, 'Cannot negate non-flag option "--$name".', '--$name');
+      _validate(
+          option.negatable!, 'Cannot negate option "--$name".', '--$name');
 
       _setFlag(_results, option, false);
     } else {
       // Walk up to the parent command if possible.
-      _validate(_parent != null, 'Could not find an option named "$name".');
+      _validate(_parent != null, 'Could not find an option named "--$name".',
+          '--$name');
       return _parent!._handleLongOption(name, value);
     }
 
@@ -307,17 +318,20 @@ class Parser {
   /// Called during parsing to validate the arguments.
   ///
   /// Throws an [ArgParserException] if [condition] is `false`.
-  void _validate(bool condition, String message) {
-    if (!condition) throw ArgParserException(message);
+  void _validate(bool condition, String message,
+      [String? args, List<String>? source, int? offset]) {
+    if (!condition) {
+      throw ArgParserException(message, null, args, source, offset);
+    }
   }
 
   /// Validates and stores [value] as the value for [option], which must not be
   /// a flag.
-  void _setOption(Map results, Option option, String value) {
+  void _setOption(Map results, Option option, String value, String arg) {
     assert(!option.isFlag);
 
     if (!option.isMultiple) {
-      _validateAllowed(option, value);
+      _validateAllowed(option, value, arg);
       results[option.name] = value;
       return;
     }
@@ -326,11 +340,11 @@ class Parser {
 
     if (option.splitCommas) {
       for (var element in value.split(',')) {
-        _validateAllowed(option, element);
+        _validateAllowed(option, element, arg);
         list.add(element);
       }
     } else {
-      _validateAllowed(option, value);
+      _validateAllowed(option, value, arg);
       list.add(value);
     }
   }
@@ -343,11 +357,11 @@ class Parser {
   }
 
   /// Validates that [value] is allowed as a value of [option].
-  void _validateAllowed(Option option, String value) {
+  void _validateAllowed(Option option, String value, String arg) {
     if (option.allowed == null) return;
 
     _validate(option.allowed!.contains(value),
-        '"$value" is not an allowed value for option "${option.name}".');
+        '"$value" is not an allowed value for option "$arg".', arg);
   }
 }