+++ CHANGELOG.md
@@ -1,5 +1,7 @@
-## 2.5.1-wip
+## 2.6.0-wip
 
+* Added source argument when throwing a `ArgParserException`.
+* Fix inconsistent `FormatException` messages
 * Require Dart 3.3
 
 ## 2.5.0